Try to insert duplicates in list of pki
The stackAlloc oracle associates a [pki] to each stack variable. What
happens when two different [pki]s are associated to the same variable?
We managed to get a segmentation fault with the executable produced
using a custom stackAlloc oracle.

The problem is probably present also in the case of parameters.